+++ srbuild/tools/compiler.py
@@ -1,14 +1,15 @@
 from srbuild.tools.flags import BuildFlags
 from srbuild.logger import G_LOGGER
 import srbuild.utils as utils
-from typing import List, Dict
+from typing import List, Union
 import abc
 
-# Responsible for translating srbuild.tools.flags.BuildFlags to actual command-line BuildFlags.
+# Responsible for translating srbuild.tools.flags.BuildFlags to actual command-line flags.
 # This class defines everything about each compiler by supplying a unified interface.
 # That means that Compiler can blindly use any CompilerDef to generate valid build commands.
 class CompilerDef(abc.ABC):
-    def executable(self) -> str:
+    @staticmethod
+    def executable() -> str:
         """
         Specifies executable associated with this CompilerDef.
         For example, this would return "clang" for Clang.
@@ -18,7 +19,8 @@ class CompilerDef(abc.ABC):
         """
         pass
 
-    def compile_only(self) -> str:
+    @staticmethod
+    def compile_only() -> str:
         """
         Specifies compile-only flag for this CompilerDef.
         For example, this would return "-c" for Clang.
@@ -28,7 +30,8 @@ class CompilerDef(abc.ABC):
         """
         pass
 
-    def output(self, path: str) -> str:
+    @staticmethod
+    def output(path: str) -> str:
         """
         Specifies command-line arguments for outputting to the specified location.
         For example, this would return "-opath" for Clang.
@@ -38,7 +41,8 @@ class CompilerDef(abc.ABC):
         """
         pass
 
-    def include(self, path: str) -> str:
+    @staticmethod
+    def include(path: str) -> str:
         """
         Specifies command-line arguments for adding an include directory.
         For example, this would return "-Ipath" for Clang.
@@ -48,7 +52,8 @@ class CompilerDef(abc.ABC):
         """
         pass
 
-    def parse_flags(self, build_flags: BuildFlags) -> List[str]:
+    @staticmethod
+    def parse_flags(build_flags: BuildFlags) -> List[str]:
         """
         Parses build flags and returns the required command-line arguments.
         """
@@ -56,16 +61,20 @@ class CompilerDef(abc.ABC):
 
 # For conventions that are common among Linux compilers.
 class LinuxCompilerDef(CompilerDef):
-    def compile_only(self) -> str:
+    @staticmethod
+    def compile_only() -> str:
         return "-c"
 
-    def output(self, path: str) -> str:
+    @staticmethod
+    def output(path: str) -> str:
         return f"-o{path}"
 
-    def include(self, path: str) -> str:
+    @staticmethod
+    def include(path: str) -> str:
         return f"-I{path}"
 
-    def parse_flags(self, build_flags: BuildFlags) -> List[str]:
+    @staticmethod
+    def parse_flags(build_flags: BuildFlags) -> List[str]:
         compiler_flags = []
         if build_flags._o:
             compiler_flags.append(f"-O{build_flags._o}")
@@ -78,16 +87,18 @@ class LinuxCompilerDef(CompilerDef):
         return compiler_flags
 
 class ClangDef(LinuxCompilerDef):
-    def executable(self) -> str:
+    @staticmethod
+    def executable() -> str:
         return "clang"
 
 class GCCDef(LinuxCompilerDef):
-    def executable(self) -> str:
+    @staticmethod
+    def executable() -> str:
         return "gcc"
 
 # Responsible for generating commands that will compile a given source file with the given flags
 class Compiler(object):
-    def __init__(self, cdef: CompilerDef):
+    def __init__(self, cdef: Union[type, CompilerDef]):
         self.cdef = cdef
 
     # The signature is everything that makes the resulting object file unique
@@ -102,12 +113,12 @@ class Compiler(object):
     # Generates the command required to compile the input file with the specified options.
     # TODO: Docstring here
     def compile(self, input_path: str, output_path: str, include_dirs: List[str]=[], flags: BuildFlags=BuildFlags()) -> List[str]:
-        includes = [self.cdef.include(dir) for dir in include_dirs]
         compiler_flags = self.cdef.parse_flags(flags)
+        includes = [self.cdef.include(dir) for dir in include_dirs]
         # The full command, including the output file and the compile-only flag.
         cmd = [self.cdef.executable(), input_path] + compiler_flags + includes + [self.cdef.compile_only(), self.cdef.output(output_path)]
         G_LOGGER.debug(f"Compile Command: {' '.join(cmd)}")
         return cmd
 
-clang = Compiler(ClangDef())
-gcc = Compiler(GCCDef())
+clang = Compiler(ClangDef)
+gcc = Compiler(GCCDef)