+++ srbuild/utils.py
@@ -17,19 +17,3 @@ def str_hash(obj) -> str:
     generated_hash = hashlib.md5(in_str.encode()).hexdigest()
     G_LOGGER.verbose(f"Generated hash {generated_hash} from '{in_str}'")
     return generated_hash
-
-
-def timestamp(path) -> int:
-    """
-    Returns the timestamp for a given path - this corresponds to the time at which the path was last modified.
-
-    Args:
-        path (str): The path to check
-
-    Returs:
-        int: The time at which the path was last modified in number of nanoseconds since the epoch, or -1 if the file does not exist.
-    """
-    try:
-        return os.path.getmtime(path)
-    except FileNotFoundError:
-        return -1